+++ app/src/main/java/com/example/gagyeboost/di/AppModules.kt
@@ -9,7 +9,7 @@ import com.example.gagyeboost.model.Repository
 import com.example.gagyeboost.model.local.AccountBookDatabase
 import com.example.gagyeboost.ui.home.AddViewModel
 import com.example.gagyeboost.ui.home.HomeViewModel
-import com.example.gagyeboost.ui.home.RecordDetailViewModel
+import com.example.gagyeboost.ui.home.detail.RecordDetailViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
