+++ app/src/main/res/layout/activity_record_detail.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="viewModel"
-            type="com.example.gagyeboost.ui.home.RecordDetailViewModel" />
+            type="com.example.gagyeboost.ui.home.detail.RecordDetailViewModel" />
 
     </data>
 
@@ -18,7 +18,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            tools:context=".ui.home.RecordDetailActivity">
+            tools:context=".ui.home.detail.RecordDetailActivity">
 
             <com.google.android.material.appbar.AppBarLayout
                 android:id="@+id/app_bar_layout_update_record"
@@ -193,11 +193,11 @@
                 app:layout_constraintTop_toBottomOf="@id/tv_address" />
 
             <TextView
-                android:id="@+id/btn_complete"
+                android:id="@+id/btn_update"
                 android:layout_width="100dp"
                 android:layout_height="wrap_content"
                 android:theme="@style/GradationButton"
-                android:text="@string/complete"
+                android:text="@string/update"
                 android:textSize="18sp"
                 android:layout_marginVertical="20dp"
                 app:layout_constraintTop_toBottomOf="@id/map"
